[BUGFIX] handler is undefined when resizing the backend
Resizing the backend browser window results in a JavaScript error.
The function that handles the re-positioning of the menus does not
work correctly. The observer requires a reference to a callback
function - currently the method was directly executed and the return
value used as function reference, which was undefined in this case.

Change-Id: Ifb2db858d7075f4fd714d737576189c0e1efc5a5
Resolves: #25787
Releases: 4.6, 4.5
Reviewed-on: http://review.typo3.org/5634
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader