[BUGFIX] Use already initialized user in OpenDocumentService
In #96318 the call to `$backendUser->start()` was removed
in OpenDocumentService. This then led to an exception, in
case a not fully initialized backend user object is used to
call `getModuleData()`.

This is fixed by no longer using an own backend user
instance, but instead accessing the already initialized
global object.

Resolves: #96330
Related: #96318
Releases: main
Change-Id: Ic887df324b2cea4eb79c48f4a50e2be28d8f0774
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72623
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>