[TASK] Unify badge rendering
Badges are currently in an inconsistent state
and sometimes the severity it shows is not
related to the contents of the badge at all.

To resolve this we are removing all auto coloring
of badges. If a badge needs a meaningful severity
it must be set explicitly.

Badges are now also not extremely rounded by default
anymore. Except for the notification in the topbar,
there is no need to have them look like notification
bubbles.

Resolves: #98311
Releases: main
Change-Id: I29dc9e8909e1c32ecc084dfaad6a3279531c389d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75679
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>