[BUGFIX] Prevent topbar button name duplication in firefox a11y tree
Firefox collects title attributes of DOM elements as
accessible names, even if their content is empty (or marked as
aria-hidden).

Since toolbar items contain icon wrapper markup that contains a title,
(which is needed for desktop mouse users, since the sibling text title
is visually hidden) and a title node as a text leaf, the title was read
out twice by screenreaders in firefox, since it's accessibility tree
collected the icon wrapper with the title as one generic node
and the text leaf as another node.

To unify browser behavior we're adding an explicit aria-hidden
attribute to the icon wrapper, in order for firefox to
interpret this DOM tree as other browser vendors do.

In addition the title attribute is removed from the search and single
clear-cache toolbar buttons in order to unify their behavior with the
other toolbar buttons which do only contain a title attribute on
the icon wrapper.

Resolves: #107723
Releases: main, 14.3
Change-Id: Iceb8243c56fafda4f610921335649d6d33b4853b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91299
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Tested-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>